feat: finished SelectActivity component
